# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PANOR Clinical Intelligence API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with role-based access
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: doctorId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentList"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment (AI suggests doctor and slot)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{id}:
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status or reschedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

  # BILLING
  /billing:
    get:
      operationId: listBills
      tags: [billing]
      summary: List billing records
      parameters:
        - name: patientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bills
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillList"
    post:
      operationId: createBill
      tags: [billing]
      summary: Generate a medical bill
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBillBody"
      responses:
        "201":
          description: Bill created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"
  /billing/{id}/pay:
    post:
      operationId: payBill
      tags: [billing]
      summary: Mark a bill as paid
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bill paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bill"

  # EPIDEMIOLOGY
...
  /ai/intake-summary:
    post:
      operationId: aiIntakeSummary
      tags: [ai]
      summary: AI-powered symptom intake summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IntakeSummaryBody"
      responses:
        "200":
          description: Clinical summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntakeSummaryResult"
  /ai/copilot:
    post:
      operationId: aiCopilot
      tags: [ai]
      summary: Real-time doctor AI co-pilot for differential diagnosis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CopilotBody"
      responses:
        "200":
          description: Copilot suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CopilotResult"
  /ai/lab-translate:
    post:
      operationId: aiLabTranslate
      tags: [ai]
      summary: Translate clinical intent to structured lab tests
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LabTranslateBody"
      responses:
        "200":
          description: Lab tests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LabTranslateResult"
  /ai/followup:
    post:
      operationId: aiFollowup
      tags: [ai]
      summary: Generate 48-hour AI follow-up assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FollowupBody"
      responses:
        "200":
          description: Follow-up result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FollowupResult"

  # DASHBOARD
...
components:
  schemas:
...
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [patient, doctor, lab, analyst, admin]
        specialization:
          type: string
          nullable: true
        medicalId:
          type: string
          nullable: true
      required: [id, email, name, role]
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        doctorId:
          type: integer
        doctorName:
          type: string
        specialty:
          type: string
        datetime:
          type: string
        duration:
          type: integer
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        priority:
          type: string
          enum: [normal, urgent]
        aiSuggested:
          type: boolean
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, patientId, patientName, doctorId, doctorName, specialty, datetime, duration, status, priority, aiSuggested, createdAt]
    AppointmentList:
      type: object
      properties:
        appointments:
          type: array
          items:
            $ref: "#/components/schemas/Appointment"
        total:
          type: integer
      required: [appointments, total]
    CreateAppointmentBody:
      type: object
      properties:
        patientId:
          type: integer
        doctorId:
          type: integer
        specialty:
          type: string
        datetime:
          type: string
        priority:
          type: string
          enum: [normal, urgent]
          default: normal
        notes:
          type: string
        symptoms:
          type: string
      required: [patientId, doctorId, specialty, datetime]
    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        datetime:
          type: string
        notes:
          type: string
    Bill:
      type: object
      properties:
        id:
          type: integer
        patientId:
          type: integer
        patientName:
          type: string
        visitId:
          type: integer
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              amount:
                type: number
              category:
                type: string
                enum: [consultation, lab, medication, other]
            required: [description, amount, category]
        total:
          type: number
        status:
          type: string
          enum: [pending, paid, insurance]
        aiSummary:
          type: string
          nullable: true
        createdAt:
          type: string
        paidAt:
          type: string
          nullable: true
      required: [id, patientId, patientName, items, total, status, createdAt]
    BillList:
      type: object
      properties:
        bills:
          type: array
          items:
            $ref: "#/components/schemas/Bill"
        total:
          type: integer
      required: [bills, total]
    CreateBillBody:
      type: object
      properties:
        patientId:
          type: integer
        visitId:
          type: integer
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              description:
                type: string
              amount:
                type: number
              category:
                type: string
                enum: [consultation, lab, medication, other]
            required: [description, amount, category]
      required: [patientId, items]
...
    IntakeSummaryBody:
      type: object
      properties:
        symptoms:
          type: string
        language:
          type: string
          enum: [en, ur]
          default: en
        patientId:
          type: integer
          nullable: true
      required: [symptoms]
    IntakeSummaryResult:
      type: object
      properties:
        chiefComplaint:
          type: string
        structuredSymptoms:
          type: array
          items:
            type: object
            properties:
              symptom:
                type: string
              duration:
                type: string
              severity:
                type: string
            required: [symptom, duration, severity]
        urgencyLevel:
          type: string
          enum: [low, moderate, high, critical]
        aiExplanation:
          type: string
        confidenceScore:
          type: number
        suggestedSpecialty:
          type: string
      required: [chiefComplaint, structuredSymptoms, urgencyLevel, aiExplanation, confidenceScore, suggestedSpecialty]
    CopilotBody:
      type: object
      properties:
        visitId:
          type: integer
        clinicalNotes:
          type: string
        patientHistory:
          type: string
        symptoms:
          type: string
      required: [visitId, clinicalNotes]
    CopilotResult:
      type: object
      properties:
        differentialDiagnoses:
          type: array
          items:
            type: object
            properties:
              diagnosis:
                type: string
              probability:
                type: number
              reasoning:
                type: string
              urgency:
                type: string
                enum: [low, moderate, high, critical]
              contributingSymptoms:
                type: array
                items:
                  type: string
            required: [diagnosis, probability, reasoning, urgency, contributingSymptoms]
        riskAlerts:
          type: array
          items:
            type: object
            properties:
              alert:
                type: string
              severity:
                type: string
              explanation:
                type: string
            required: [alert, severity, explanation]
        recommendedActions:
          type: array
          items:
            type: string
        confidenceScore:
          type: number
        aiExplanation:
          type: string
      required: [differentialDiagnoses, riskAlerts, recommendedActions, confidenceScore, aiExplanation]
    LabTranslateBody:
      type: object
      properties:
        clinicalIntent:
          type: string
        patientContext:
          type: string
          nullable: true
        priority:
          type: string
          enum: [routine, urgent, stat]
          default: routine
      required: [clinicalIntent]
    LabTranslateResult:
      type: object
      properties:
        tests:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              code:
                type: string
              priority:
                type: string
                enum: [routine, urgent, stat]
              instructions:
                type: string
              rationale:
                type: string
            required: [name, code, priority, instructions, rationale]
        aiExplanation:
          type: string
        totalEstimatedTime:
          type: string
      required: [tests, aiExplanation, totalEstimatedTime]
    FollowupBody:
      type: object
      properties:
        visitId:
          type: integer
        patientId:
          type: integer
        patientResponse:
          type: string
      required: [visitId, patientId, patientResponse]
    FollowupResult:
      type: object
      properties:
        status:
          type: string
          enum: [improving, stable, deteriorating]
        assessment:
          type: string
        recommendations:
          type: array
          items:
            type: string
        requiresUrgentAttention:
          type: boolean
        confidenceScore:
          type: number
      required: [status, assessment, recommendations, requiresUrgentAttention, confidenceScore]
...