# lib/api-zod/src/generated/types/followupResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PANOR Clinical Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { FollowupResultStatus } from "./followupResultStatus";

export interface FollowupResult {
  status: FollowupResultStatus;
  assessment: string;
  recommendations: string[];
  requiresUrgentAttention: boolean;
  confidenceScore: number;
}