# lib/api-zod/src/generated/types/labTranslateResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PANOR Clinical Intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { LabTranslateResultTestsItem } from "./labTranslateResultTestsItem";

export interface LabTranslateResult {
  tests: LabTranslateResultTestsItem[];
  aiExplanation: string;
  totalEstimatedTime: string;
}